Move dtone package to to airtime/dtone
