+++ providers/airtime/dtone/client_test.go
@@ -6,7 +6,7 @@ import (
 	"net/http/httptest"
 	"testing"
 
-	"github.com/nyaruka/goflow/providers/dtone"
+	"github.com/nyaruka/goflow/providers/airtime/dtone"
 
 	"github.com/stretchr/testify/assert"
 )
