+++ providers/airtime/dtone/utils_test.go
@@ -4,7 +4,7 @@ import (
 	"encoding/json"
 	"testing"
 
-	"github.com/nyaruka/goflow/providers/dtone"
+	"github.com/nyaruka/goflow/providers/airtime/dtone"
 
 	"github.com/shopspring/decimal"
 	"github.com/stretchr/testify/assert"
